% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parappend.R
\name{parappend}
\alias{parappend}
\title{Append data to a file with parallel safety}
\usage{
parappend(
  out.file,
  data,
  append = TRUE,
  col.names = FALSE,
  sep = "\\t",
  quote = FALSE,
  row.names = FALSE,
  na = "NA",
  rm.lock = FALSE,
  max.retries = Inf,
  time.sleep = 0.01,
  ...
)
}
\arguments{
\item{out.file}{The output file path}

\item{data}{The data to append (eg. a data.frame or data.table)}

\item{append}{Whether to append to the file (default: TRUE) or overwrite (FALSE)}

\item{col.names}{Whether to write column names (default: FALSE), defaults to TRUE if file does not exist.}

\item{sep}{The separator to use (default: "\\t")}

\item{quote}{Whether to quote character/factor columns (default: FALSE)}

\item{row.names}{Whether to write row names (default: FALSE)}

\item{na}{The string to use for missing values (default: "NA")}

\item{rm.lock}{Whether to remove the lock file after releasing the lock (default: FALSE)}

\item{max.retries}{Maximum number of retries to acquire lock (default: Inf)}

\item{time.sleep}{Time to wait (in seconds) between retries (default: 0.01)}

\item{...}{Additional arguments passed to data.table::fwrite}
}
\value{
invisible()
}
\description{
Append data to a file with parallel safety
}
\examples{
\dontrun{
df <- data.frame(x = 1:10, y = letters[1:10])
if(file.exists("parappend.test.tsv")) unlink("parappend.test.tsv")
myfun <- function(i) {
  Sys.sleep(runif(1, 0, 0.1))  # Simulate variable processing time
  parappend::parappend("parappend.test.tsv", df[i, , drop = FALSE])
}
tmp <- parallel::mclapply(seq_len(nrow(df)), myfun, mc.cores = 4)
# Verify the output
result <- data.table::fread("parappend.test.tsv")
print(result)
# Clean up
unlink("parappend.test.tsv")
unlink("parappend.test.tsv.lock")

}
}
\seealso{
\code{\link[data.table]{fwrite}}
}
